% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/S4classes.R
\docType{class}
\name{AnnoResource-class}
\alias{AnnoResource-class}
\title{document the provenance of a Bioconductor annotation resource like GO}
\description{
document the provenance of a Bioconductor annotation resource like GO
}
\section{Slots}{

\describe{
\item{\code{declaredDate}}{Date instance derived from source, usually parsing content to find the date}

\item{\code{retrievedData}}{Date instance specified for retrieval event}

\item{\code{sourceURL}}{character(1) full URL}
}}

