% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_modification_time.R
\name{chk_modification_time}
\alias{chk_modification_time}
\title{check modification time for content at a URL}
\usage{
chk_modification_time(url)
}
\arguments{
\item{url}{string}
}
\description{
check modification time for content at a URL
}
\examples{
check_modification_time("http://current.geneontology.org/ontology/go-basic.obo")
}
