% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download_and_gzip.R
\name{download_and_gzip}
\alias{download_and_gzip}
\title{use curl to retrieve and gzfile to compress resources}
\usage{
download_and_gzip(url, dest_gz)
}
\arguments{
\item{url}{string}

\item{dest_gz}{a filename to create, with .gz suffix}
}
\value{
invisibly returns path to created compressed file
}
\description{
use curl to retrieve and gzfile to compress resources
}
\examples{
\dontrun{
download_and_gzip("http://current.geneontology.org/ontology/go-basic.obo", "./go-basic.obo.gz")
}
}
