% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GO_utils.R
\name{get_go_date}
\alias{get_go_date}
\title{obtain value of `data-version` in go-basic.obo}
\usage{
get_go_date(path)
}
\arguments{
\item{path}{character string}
}
\description{
obtain value of `data-version` in go-basic.obo
}
