% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/S4classes.R
\name{getGOobo}
\alias{getGOobo}
\title{retrieve GO in OBO format}
\usage{
getGOobo(
  gourl = "http://current.geneontology.org/ontology/go-basic.obo",
  destdir
)
}
\arguments{
\item{gourl}{string}

\item{destdir}{path}
}
\description{
retrieve GO in OBO format
}
\note{
curl is used to retrieve go-basic.obo by default, will not overwrite existing file
}
