% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/S4classes.R
\name{resourceDate,AnnoResource-method}
\alias{resourceDate,AnnoResource-method}
\title{extract date of resource}
\usage{
\S4method{resourceDate}{AnnoResource}(x)
}
\arguments{
\item{x}{instance of AnnoResource}
}
\description{
extract date of resource
}
