% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parseOBO.R
\name{parseOBO}
\alias{parseOBO}
\title{package-resident version of parseOBO.R utility from pipeline}
\usage{
parseOBO(goresource, workdir)
}
\arguments{
\item{goresource}{instance of BiocGO}

\item{workdir}{path files will be written here}
}
\description{
package-resident version of parseOBO.R utility from pipeline
}
\note{
Writes graph_path.txt, term2term.txt term_definition.txt, term_synonym.txt and term.txt in workdir.
Dec 2025: Code is modified to add colnames to these txt files.  This will facilitate use of
dbWriteTable after ingesting the file, eliminating the need for scripting sqlite .read commands.
}
